Found Minor Bug
When building the src folder is being copied to the build folder.  However, this compiled js file should be placed in the build folder and not a src folder inside build.

Commiting to keep an working image while working on this.

The static zip file contains the new build for the front-end.