Minor Updates
Solved the issue about the TS compiling.
    - When building the js files, the source js files were added to ./build/src/*. This was
      not ideal as the directory should be "./build/*" . The issue that was happening is that when building, the process was looking for the static js files located in  "./www". All of the source ts file are located in "./src"  however since the building process was looking for the static js files, it copied the src and www folders to build.

Fix:   Exclude the "./www" js files in tsconfig.json, and set  "rooDir": "./src".
With this, the built js file are now placed in "./build" without the src folder.
On a previous version of nestJS , 6.13.0, this was not necessary. This issue happened after updating nestJS to 7.2.0.